fix(E27): resolve Devin Review issues - architecture boundary, RLS, SYSTEM_MAP sync
Addresses all 10 Devin Review findings:

CRITICAL FIXES (🔴):
1. Architecture boundary violation: moved ReflectionRecord from adapters to core/models
   - Moved ReflectionRecord from adapters/storage/postgres_reflection_models.py
     to core/models/reflection.py (domain model, not adapter-specific)
   - Updated all imports across 6 files (ports, adapters, tests)
   - Deleted obsolete postgres_reflection_models.py

2. SYSTEM_MAP.md §1.5 and §2.2 updates:
   - Added InMemoryReflectionStore to §1.5 Adapters
   - Added reflection_persistence_helper.py to §1.5 Adapters
   - Added InMemoryReflectionStore → ReflectionStore mapping to §2.2

3. SYSTEM_MAP-ru.md full sync:
   - Synced §1.1 (added ReflectionRecord to core/models/reflection.py)
   - Synced §1.2 (added core/ports/reflection_store.py)
   - Synced §1.5 (added E27 adapters)
   - Synced §2.2 (added adapter↔port mapping)

WARNINGS FIXED (🟡):
4. Markdown table syntax: fixed '||' → '|' in SYSTEM_MAP.md §1.2

5. InMemoryReflectionStore.get() RLS enforcement:
   - Added RLS filter to get() method (was incorrectly bypassing it)
   - Now matches PostgreSQL RLS behavior: USING policy applies to all SELECT queries

TYPE SAFETY:
6. Added 'assert record.id is not None' before store.get(record.id) calls
   in test_reflection_persistence_integration.py (2 locations)

All checks passing:
- ruff check: ✓
- ruff format: ✓
- pyright: ✓ (0 errors)
- bandit: ✓
- pytest: ✓ (92.38% coverage, 680 passed)
  - 2 pre-existing failures unrelated to E27 (deploy zip hash, demo script)

Refs: Devin Review on PR #412
