--- src/atman/adapters/storage/postgres_reflection_models.py
@@ -1,137 +1,0 @@
-"""
-PostgreSQL persistence models for Reflection Engine.
-
-These models map directly to the `public.reflections` table schema.
-They are used by the PostgreSQL ReflectionStore adapter.
-"""
-
-from datetime import UTC, datetime
-from typing import Any
-from uuid import UUID
-
-from pydantic import BaseModel, ConfigDict, Field, field_validator
-
-from atman.core.models.reflection import ReflectionLevel
-
-
-class ReflectionRecord(BaseModel):
-    """
-    PostgreSQL persistence model for reflections table.
-
-    Maps to public.reflections schema. This is the canonical persistent format
-    for reflections (micro/daily/deep) stored in PostgreSQL.
-
-    Schema corresponds to:
-    - id: BIGSERIAL PRIMARY KEY
-    - agent_id: UUID NOT NULL REFERENCES agents(id)
-    - level: reflection_level NOT NULL (enum: micro, daily, deep)
-    - created_at: TIMESTAMPTZ NOT NULL DEFAULT now()
-    - session_id: UUID REFERENCES sessions(id) -- populated for level='micro'
-    - period_start: TIMESTAMPTZ -- populated for level='daily'/'deep'
-    - period_end: TIMESTAMPTZ -- populated for level='daily'/'deep'
-    - content: TEXT NOT NULL -- the reflection itself (generated by LLM)
-    - summary: TEXT -- short title/summary
-    - experience_refs: UUID[] NOT NULL DEFAULT '{}' -- experiences analyzed
-    - reframing_note_ids: UUID[] NOT NULL DEFAULT '{}' -- reframing notes created
-    - model_provider: TEXT -- 'ollama', 'anthropic', etc.
-    - model_name: TEXT -- e.g. 'qwen3:14b'
-    - schema_version: INTEGER NOT NULL DEFAULT 1
-    - metadata: JSONB NOT NULL DEFAULT '{}'
-    """
-
-    # Primary key (BIGSERIAL in Postgres, int in Python after retrieval)
-    id: int | None = Field(default=None, description="Database ID (BIGSERIAL); None before INSERT")
-
-    # Core fields
-    agent_id: UUID = Field(description="Agent this reflection belongs to")
-    level: ReflectionLevel = Field(description="Reflection level: micro/daily/deep")
-    created_at: datetime = Field(
-        default_factory=lambda: datetime.now(UTC),
-        description="When this reflection was created",
-    )
-
-    # Session context (for micro reflections)
-    session_id: UUID | None = Field(
-        default=None,
-        description="Session ID (populated only for level='micro', NULL for daily/deep)",
-    )
-
-    # Time period context (for daily/deep reflections)
-    period_start: datetime | None = Field(
-        default=None,
-        description="Start of period analyzed (populated for level='daily'/'deep', NULL for micro)",
-    )
-    period_end: datetime | None = Field(
-        default=None,
-        description="End of period analyzed (populated for level='daily'/'deep', NULL for micro)",
-    )
-
-    # Content
-    content: str = Field(
-        min_length=1, description="The reflection text itself (generated by reflection model)"
-    )
-    summary: str | None = Field(default=None, description="Optional short title/summary")
-
-    # References
-    experience_refs: list[UUID] = Field(
-        default_factory=list, description="UUIDs of experiences analyzed in this reflection"
-    )
-    reframing_note_ids: list[UUID] = Field(
-        default_factory=list,
-        description="UUIDs of reframing_notes created by this reflection",
-    )
-
-    # Model metadata (for traceability)
-    model_provider: str | None = Field(
-        default=None, description="LLM provider: 'ollama', 'anthropic', etc."
-    )
-    model_name: str | None = Field(default=None, description="Specific model: e.g. 'qwen3:14b'")
-
-    # Schema versioning and metadata
-    schema_version: int = Field(
-        default=1, ge=1, description="Schema version for migrations (default 1)"
-    )
-    metadata: dict[str, Any] = Field(
-        default_factory=dict, description="Additional metadata (JSONB in PostgreSQL)"
-    )
-
-    @field_validator("content")
-    @classmethod
-    def validate_content_not_empty(cls, v: str) -> str:
-        """Ensure content is not empty or whitespace-only."""
-        if not v or not v.strip():
-            raise ValueError("content cannot be empty")
-        return v.strip()
-
-    @field_validator("schema_version")
-    @classmethod
-    def validate_schema_version_positive(cls, v: int) -> int:
-        """Ensure schema_version is positive."""
-        if v < 1:
-            raise ValueError("schema_version must be >= 1")
-        return v
-
-    @field_validator("experience_refs", "reframing_note_ids")
-    @classmethod
-    def validate_uuid_lists(cls, v: list[UUID]) -> list[UUID]:
-        """Ensure UUID lists contain valid UUIDs."""
-        return list(v)  # Pydantic will validate UUID type automatically
-
-    model_config = ConfigDict(
-        validate_assignment=True,
-        json_schema_extra={
-            "example": {
-                "agent_id": "123e4567-e89b-12d3-a456-426614174000",
-                "level": "micro",
-                "session_id": "123e4567-e89b-12d3-a456-426614174001",
-                "content": "Session went well. User was collaborative, I maintained clarity.",
-                "summary": "Positive collaborative session",
-                "experience_refs": [],
-                "reframing_note_ids": [],
-                "model_provider": "ollama",
-                "model_name": "qwen3:14b",
-                "schema_version": 1,
-                "metadata": {},
-            }
-        },
-    )
